% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asi.R
\docType{data}
\name{asi}
\alias{asi}
\title{Did patient use intravenous drugs}
\format{
A tibble with 3,560 rows and
2 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}

\item{\strong{used_iv}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = No) \cr
    \tab  \cr
   \emph{Description:} \tab Self-reported history of IV drug use \cr
    \tab  \cr
   \emph{Levels:} \tab No, Yes \cr
}


}

}
}
\usage{
data(asi)
}
\description{
IV drug use information from the Addiction Severity Index
}
\keyword{datasets}
