% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detox.R
\docType{data}
\name{detox}
\alias{detox}
\title{Start and Stop of Detox}
\format{
A tibble with 1,316 rows and 3
variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{what}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = admission) \cr
    \tab  \cr
   \emph{Description:} \tab Indicator for start or stop of detox \cr
    \tab  \cr
   \emph{Levels:} \tab admission, discharge \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Day of start or stop of detox \cr
}


}
}
}
\usage{
data(detox)
}
\description{
This is the start and stop date for detox (if known).
}
\keyword{datasets}
