% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everybody.R
\docType{data}
\name{everybody}
\alias{everybody}
\title{Everybody with any data}
\format{
A tibble with 3,560 rows and
2 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{project}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = 27) \cr
    \tab  \cr
   \emph{Description:} \tab CTN project number \cr
    \tab  \cr
   \emph{Levels:} \tab 27, 30, 51 \cr
}


}
}
}
\usage{
data(everybody)
}
\description{
This is a list of each person with their original study project.
}
\keyword{datasets}
