% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagerstrom.R
\docType{data}
\name{fagerstrom}
\alias{fagerstrom}
\title{Fagerstrom Test for Nicotine Dependence}
\format{
A tibble with 3,119 rows and
4 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{is_smoker}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab No = Is not a smoker, Yes = Is a smoker \cr
    \tab  \cr
   \emph{Levels:} \tab No, Yes \cr
}


}
\item{\strong{ftnd}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Fagerstrom Test For Nicotine Dependence Score 0-10 \cr
    \tab  \cr
   \emph{Levels:} \tab 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10 \cr
}


}
\item{\strong{per_day}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Cigarettes per day \cr
    \tab  \cr
   \emph{Levels:} \tab , 10 OR LESS, 11-20, 21-30, 31 OR MORE \cr
}


}
}
}
\usage{
fagerstrom
}
\description{
Information on the intensity of physical addiction to nicotine at baseline. See the vignette
\href{../doc/harmonization.html}{Harmonization Information} for more details.
}
\keyword{datasets}
