% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_survey.R
\docType{data}
\name{first_survey}
\alias{first_survey}
\title{First Survey Date}
\format{
A tibble with 3,453 rows and
2 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Study day \cr
}


}
}
}
\usage{
data(first_survey)
}
\description{
This file contains the dates for the demographics survey in
CTN-0027 and CTN-0030 and the medical-psychiatric history for CTN-0051.
}
\keyword{datasets}
