% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pain.R
\docType{data}
\name{pain}
\alias{pain}
\title{Self-Reported Pain}
\format{
A tibble with 3,082 rows and
3 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{pain}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = Missing) \cr
    \tab  \cr
   \emph{Description:} \tab Pain severity \cr
    \tab  \cr
   \emph{Levels:} \tab Missing, No Pain, Severe Pain, Very mild to Moderate Pain \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Study day \cr
}


}
}
}
\usage{
data(pain)
}
\description{
This is self-reported pain from the SF-36 (ctn_0027 and
ctn_0030) and EuroQoL (ctn_0051). See the
\href{../doc/harmonization.html#pain}{Harmonization Information} vignette for
more details.
}
\keyword{datasets}
