% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\docType{data}
\name{randomization}
\alias{randomization}
\title{Randomization Data}
\format{
A tibble with 4,691 rows and
4 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{treatment}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = Inpatient BUP) \cr
    \tab  \cr
   \emph{Description:} \tab What treatment is prescribed? \cr
    \tab  \cr
   \emph{Levels:} \tab Inpatient BUP, Inpatient NR-NTX, Methadone, Outpatient BUP, Outpatient BUP + EMM, Outpatient BUP + SMM \cr
}


}
\item{\strong{which}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = 1) \cr
    \tab  \cr
   \emph{Description:} \tab Indicator of which randomization. Needed because CTN 30 has two randomization dates. \cr
    \tab  \cr
   \emph{Levels:} \tab 1, 2 \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Study day of treatment entry \cr
}


}
}
}
\usage{
data(randomization)
}
\description{
This is the information on the treatment group. Note that CTN30
had two randomization events.
}
\keyword{datasets}
