% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening_date.R
\docType{data}
\name{screening_date}
\alias{screening_date}
\title{Screening Date Information}
\format{
A tibble with 3,430 rows and
3 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{screening_day}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab The day of screening \cr
}


}
\item{\strong{day_zero}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Best guess at the end of TLFB \cr
}


}
}
}
\usage{
data(screening_date)
}
\description{
The information on the screening data and baseline drug
screening data is complex and inconsistent across studies. This file has
information on dates around the screening visit and baseline timeline
follow back assessments.
}
\keyword{datasets}
