% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment.R
\docType{data}
\name{treatment}
\alias{treatment}
\title{Amount of Study Drug Per Day}
\format{
A tibble with 216,242 rows and
3 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{amount}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab The amount of drugs received on a day. Value is 1 for injections and mg otherwise \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Study day \cr
}


}
}
}
\usage{
data(treatment)
}
\description{
The doses (usually in milligrams) of the study drug administered
to each subject by day. When the study drug is listed as an injection, then
the amount is recorded as 1.
}
\keyword{datasets}
