% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uds_temp.R
\docType{data}
\name{uds_temp}
\alias{uds_temp}
\title{Urine Drug Screening (UDS) Temperature}
\format{
A tibble with 36,680 rows and
3 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{was_temp_ok}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = 0) \cr
    \tab  \cr
   \emph{Description:} \tab was the temperature 92째F \eqn{\le X \le} 96째F OR 33.3째C \eqn{\le X \le} 35.5째C? 0 = "No", 1 = "Yes", 99 = "Unknown" \cr
    \tab  \cr
   \emph{Levels:} \tab 0, 1, 99 \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Study day \cr
}


}
}
}
\usage{
data(uds_temp)
}
\description{
This is information on whether the urine temperature was in the
acceptable range.
}
\keyword{datasets}
