% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withdrawal_pre_post.R
\docType{data}
\name{withdrawal_pre_post}
\alias{withdrawal_pre_post}
\title{Patient Withdrawal Symptoms Pre and Post Induction}
\format{
A tibble with 4,805 rows and
4 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{what}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = post) \cr
    \tab  \cr
   \emph{Description:} \tab Indicator of induction day type: "pre" or "post" \cr
    \tab  \cr
   \emph{Levels:} \tab post, pre \cr
}


}
\item{\strong{withdrawal}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = 0) \cr
    \tab  \cr
   \emph{Description:} \tab \if{html}{\out{<ul>}}\if{html}{\out{<li>}}0 = "None"\if{html}{\out{</li>}}\if{html}{\out{<li>}}1 = "mild"\if{html}{\out{</li>}}\if{html}{\out{<li>}}2 = "moderate"\if{html}{\out{</li>}}\if{html}{\out{<li>}}3 = "severe"\if{html}{\out{</li>}}\if{html}{\out{</ul>}} \cr
    \tab  \cr
   \emph{Levels:} \tab 0, 1, 2, 3 \cr
}


}
\item{\strong{when}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Day of assessment \cr
}


}
}
}
\usage{
data(withdrawal_pre_post)
}
\description{
This is a information on the severity of withdrawal symptoms.
}
\keyword{datasets}
