% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qol.R
\docType{data}
\name{qol}
\alias{qol}
\title{Quality of Life}
\format{
A tibble with 657 rows and
2 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{is_homeless}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = No) \cr
    \tab  \cr
   \emph{Description:} \tab Are you currently homeless or living in a shelter? \cr
    \tab  \cr
   \emph{Levels:} \tab No, Yes \cr
}


}
}
}
\usage{
data(qol)
}
\description{
This is quality of life data from the PhenX Quality of Life
survey (see \url{https://www.phenxtoolkit.org/protocols/view/221302}).
This was used by the Clinical Trials Network protocol CTN-0051.
}
\keyword{datasets}
