% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_masked.R
\docType{data}
\name{site_masked}
\alias{site_masked}
\title{Site Regrouped}
\format{
A tibble with 3,560 rows and
2 variables:

\describe{
\item{\strong{who}}{\tabular{ll}{
   \emph{Type:} \tab integer \cr
    \tab  \cr
   \emph{Description:} \tab Patient ID \cr
}


}
\item{\strong{site_masked}}{\tabular{ll}{
   \emph{Type:} \tab factor (First/Reference level = 2) \cr
    \tab  \cr
   \emph{Description:} \tab Study Site Regrouped \cr
    \tab  \cr
   \emph{Levels:} \tab 2, 5, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36 \cr
}


}
}
}
\usage{
data(site_masked)
}
\description{
This is study site (clinic/research site number) information.
Large study sites were split into smaller groups and small study site were
grouped into larger sites. The regrouped study sites are each approximately
100 subjects. This was done to protect the anonymity of the study sites.
}
\keyword{datasets}
